requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic==2.5.1
pydantic-settings==2.1.0
boto3==1.29.3
asyncpg==0.29.0
sqlalchemy==2.0.23
alembic==1.12.1
langchain==0.0.339
pinecone-client==2.2.4
openai==1.3.3
python-jose==3.3.0
passlib==1.7.4
redis==5.0.1
prometheus-client==0.19.0
celery==5.3.4
httpx==0.25.1
python-multipart==0.0.6
psycopg2-binary==2.9.9
email-validator==2.1.0.post1
